# docs/openapi.json
{
  "info": {
    "description": "REST API extension for Burp Suite Professional that enables programmatic access to proxy traffic, scanner functionality, scope management, and collaborative testing workflows. Features advanced response analysis, real-time proxy interception, and enhanced security testing capabilities using Montoya API 2025.8+.",
    "title": "burp2api - Burp Suite REST API Extension",
    "version": "1.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/bchecks": {
      "get": {
        "description": "List imported BChecks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List imported BChecks",
        "tags": [
          "BChecks"
        ]
      },
      "post": {
        "description": "Import a BCheck from an inline script (requires: script; optional: name, description, enabled)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Import a BCheck from an inline script (requires: script; optional: name, description, enabled)",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/import-file": {
      "post": {
        "description": "Import a BCheck from a file on the Burp host (requires: file_path; optional: name, description, enabled)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Import a BCheck from a file on the Burp host (requires: file_path; optional: name, description, enabled)",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/security-rules": {
      "get": {
        "description": "The rules POST /bchecks/validate enforces against a script",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "The rules POST /bchecks/validate enforces against a script",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/stats": {
      "get": {
        "description": "BCheck counts and status summary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "BCheck counts and status summary",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/templates": {
      "get": {
        "description": "Starter BCheck script templates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Starter BCheck script templates",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/validate": {
      "post": {
        "description": "Security-validate a BCheck script without importing it (requires: script)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Security-validate a BCheck script without importing it (requires: script)",
        "tags": [
          "BChecks"
        ]
      }
    },
    "/bchecks/{bcheckId}": {
      "delete": {
        "description": "Delete a BCheck (requires: bcheckId path parameter)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete a BCheck (requires: bcheckId path parameter)",
        "tags": [
          "BChecks"
        ]
      },
      "get": {
        "description": "Get one BCheck's details (requires: bcheckId path parameter)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get one BCheck's details (requires: bcheckId path parameter)",
        "tags": [
          "BChecks"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local Burp API Server",
      "url": "http://localhost:7850"
    }
  ],
...
}